Updated clang-format style pass

+++ source/engine/common/stringtable.cpp
@@ -21,11 +21,9 @@
 //
 //-----------------------------------------------------------------------------
 
-
 #include "odamex.h"
 
 #include "stringtable.h"
-
 
 #include "cmdlib.h"
 #include "i_system.h"
@@ -36,274 +34,272 @@
 /**
  * @brief Map a ZDoom game name to Odamex's internals and returns true if
  *        the current game is the passed string.
- * 
+ *
  * @param str String to check against.
  * @return True if the game matches the passed string, otherwise false.
  */
-static bool IfGameZDoom(const std::string& str)
+static bool IfGameZDoom(const std::string &str)
 {
-	if (!stricmp(str.c_str(), "doom") && ::gamemode != retail_chex &&
-	    ::gamemode != undetermined)
-	{
-		return true;
-	}
+    if (!stricmp(str.c_str(), "doom") && ::gamemode != retail_chex && ::gamemode != undetermined)
+    {
+        return true;
+    }
 
-	if (!stricmp(str.c_str(), "chex") && ::gamemode == retail_chex)
-	{
-		return true;
-	}
+    if (!stricmp(str.c_str(), "chex") && ::gamemode == retail_chex)
+    {
+        return true;
+    }
 
-	// We don't support anything else.
-	return false;
+    // We don't support anything else.
+    return false;
 }
 
-bool StringTable::canSetPassString(int pass, const std::string& name) const
+bool StringTable::canSetPassString(int pass, const std::string &name) const
 {
-	StringHash::const_iterator it = _stringHash.find(name);
+    StringHash::const_iterator it = _stringHash.find(name);
 
-	// New string?
-	if (it == _stringHash.end())
-		return true;
+    // New string?
+    if (it == _stringHash.end())
+        return true;
 
-	// Found an entry, does the string exist?
-	if ((*it).second.string.first == false)
-		return true;
+    // Found an entry, does the string exist?
+    if ((*it).second.string.first == false)
+        return true;
 
-	// Was the string set with a less exact pass?
-	if ((*it).second.pass >= pass)
-		return true;
+    // Was the string set with a less exact pass?
+    if ((*it).second.pass >= pass)
+        return true;
 
-	return false;
+    return false;
 }
 
 void StringTable::clearStrings()
 {
-	_stringHash.clear();
+    _stringHash.clear();
 }
 
 //
 // Loads a language
 //
-void StringTable::loadLanguage(const char* code, bool exactMatch, int pass, char* lump,
-                               size_t lumpLen)
+void StringTable::loadLanguage(const char *code, bool exactMatch, int pass, char *lump, size_t lumpLen)
 {
-	OScannerConfig config = {
-	    "LANGUAGE", // lumpName
-	    false,      // semiComments
-	    true,       // cComments
-	};
-	OScanner os = OScanner::openBuffer(config, lump, lump + lumpLen);
-	while (os.scan())
-	{
-		// Parse a language section.
-		bool shouldParseSection = false;
+    OScannerConfig config = {
+        "LANGUAGE", // lumpName
+        false,      // semiComments
+        true,       // cComments
+    };
+    OScanner os = OScanner::openBuffer(config, lump, lump + lumpLen);
+    while (os.scan())
+    {
+        // Parse a language section.
+        bool shouldParseSection = false;
 
-		os.assertTokenIs("[");
-		while (os.scan())
-		{
-			// Code to check against.
-			char checkCode[4] = {'\0', '\0', '\0', '\0'};
+        os.assertTokenIs("[");
+        while (os.scan())
+        {
+            // Code to check against.
+            char checkCode[4] = {'\0', '\0', '\0', '\0'};
 
-			if (os.compareToken("]"))
-			{
-				break;
-			}
-			else if (os.compareToken("default"))
-			{
-				// Default has a speical ID.
-				strncpy(checkCode, "**", 3);
-			}
-			else
-			{
-				// Turn the language into an ID.
-				const std::string& lang = os.getToken();
+            if (os.compareToken("]"))
+            {
+                break;
+            }
+            else if (os.compareToken("default"))
+            {
+                // Default has a speical ID.
+                strncpy(checkCode, "**", 3);
+            }
+            else
+            {
+                // Turn the language into an ID.
+                const std::string &lang = os.getToken();
 
-				if (lang.length() == 2 || lang.length() == 3)
-				{
-					strncpy(checkCode, lang.c_str(), lang.length());
-				}
-				else
-				{
-					os.error("Language identifier must be 2 or 3 characters");
-				}
-			}
+                if (lang.length() == 2 || lang.length() == 3)
+                {
+                    strncpy(checkCode, lang.c_str(), lang.length());
+                }
+                else
+                {
+                    os.error("Language identifier must be 2 or 3 characters");
+                }
+            }
 
-			if (exactMatch && strncmp(code, checkCode, 3) == 0)
-			{
-				shouldParseSection = true;
-			}
-			else if (!exactMatch && strncmp(code, checkCode, 2) == 0)
-			{
-				shouldParseSection = true;
-			}
-		}
+            if (exactMatch && strncmp(code, checkCode, 3) == 0)
+            {
+                shouldParseSection = true;
+            }
+            else if (!exactMatch && strncmp(code, checkCode, 2) == 0)
+            {
+                shouldParseSection = true;
+            }
+        }
 
-		if (shouldParseSection)
-		{
-			// Parse all of the strings in this section.
-			while (os.scan())
-			{
-				if (os.compareToken("["))
-				{
-					// We reached the end of the section.
-					os.unScan();
-					break;
-				}
+        if (shouldParseSection)
+        {
+            // Parse all of the strings in this section.
+            while (os.scan())
+            {
+                if (os.compareToken("["))
+                {
+                    // We reached the end of the section.
+                    os.unScan();
+                    break;
+                }
 
-				// $ifgame() does not appear to be documented in the wiki,
-				// but it causes the next string to only be set it the game
-				// matches up.
-				bool skip = false;
-				if (os.compareToken("$"))
-				{
-					os.scan();
-					os.assertTokenIs("ifgame");
-					os.scan();
-					os.assertTokenIs("(");
-					os.scan();
-					skip = !IfGameZDoom(os.getToken());
-					os.scan();
-					os.assertTokenIs(")");
-					os.scan();
-				}
+                // $ifgame() does not appear to be documented in the wiki,
+                // but it causes the next string to only be set it the game
+                // matches up.
+                bool skip = false;
+                if (os.compareToken("$"))
+                {
+                    os.scan();
+                    os.assertTokenIs("ifgame");
+                    os.scan();
+                    os.assertTokenIs("(");
+                    os.scan();
+                    skip = !IfGameZDoom(os.getToken());
+                    os.scan();
+                    os.assertTokenIs(")");
+                    os.scan();
+                }
 
-				// String name
-				const std::string& name = os.getToken();
+                // String name
+                const std::string &name = os.getToken();
 
-				// If we can find the token, skip past the string
-				if (!canSetPassString(pass, name))
-				{
-					while (os.scan())
-					{
-						if (os.compareToken(";"))
-							break;
-					}
-					continue;
-				}
+                // If we can find the token, skip past the string
+                if (!canSetPassString(pass, name))
+                {
+                    while (os.scan())
+                    {
+                        if (os.compareToken(";"))
+                            break;
+                    }
+                    continue;
+                }
 
-				os.scan();
-				os.assertTokenIs("=");
+                os.scan();
+                os.assertTokenIs("=");
 
-				// Grab the string value.
-				std::string value;
-				while (os.scan())
-				{
-					const std::string piece = os.getToken();
-					if (piece.compare(";") == 0)
-					{
-						// Found the end of the string, next batter up.
-						break;
-					}
+                // Grab the string value.
+                std::string value;
+                while (os.scan())
+                {
+                    const std::string piece = os.getToken();
+                    if (piece.compare(";") == 0)
+                    {
+                        // Found the end of the string, next batter up.
+                        break;
+                    }
 
-					value += piece;
-				}
+                    value += piece;
+                }
 
-				replaceEscapes(value);
-				if (skip)
-				{
-					continue;
-				}
-				setPassString(pass, name, value);
-			}
-		}
-		else
-		{
-			// Skip past all of the strings in this section.
-			while (os.scan())
-			{
-				if (os.compareToken("["))
-				{
-					// Found another section, parse it.
-					os.unScan();
-					break;
-				}
-			}
-		}
-	}
+                replaceEscapes(value);
+                if (skip)
+                {
+                    continue;
+                }
+                setPassString(pass, name, value);
+            }
+        }
+        else
+        {
+            // Skip past all of the strings in this section.
+            while (os.scan())
+            {
+                if (os.compareToken("["))
+                {
+                    // Found another section, parse it.
+                    os.unScan();
+                    break;
+                }
+            }
+        }
+    }
 }
 
-void StringTable::loadStringsLump(const int lump, const char* lumpname, const bool engOnly)
+void StringTable::loadStringsLump(const int lump, const char *lumpname, const bool engOnly)
 {
-	// Can't use Z_Malloc this early, so we use raw new/delete.
-	size_t len = W_LumpLength(lump);
-	char* languageLump = new char[len + 1];
-	W_ReadLump(lump, languageLump);
-	languageLump[len] = '\0';
+    // Can't use Z_Malloc this early, so we use raw new/delete.
+    size_t len          = W_LumpLength(lump);
+    char  *languageLump = new char[len + 1];
+    W_ReadLump(lump, languageLump);
+    languageLump[len] = '\0';
 
-	// String replacement pass.  Strings in an later pass can be replaced
-	// by a string in an earlier pass from another lump.
-	int pass = 1;
+    // String replacement pass.  Strings in an later pass can be replaced
+    // by a string in an earlier pass from another lump.
+    int pass = 1;
 
-	if (!engOnly)
-	{
-		// Load language-specific strings.
-		for (size_t i = 0; i < ARRAY_LENGTH(::LanguageIDs); i++)
-		{
-			// Deconstruct code into something less confusing.
-			char code[4];
-			UNMAKE_ID(code, ::LanguageIDs[i]);
+    if (!engOnly)
+    {
+        // Load language-specific strings.
+        for (size_t i = 0; i < ARRAY_LENGTH(::LanguageIDs); i++)
+        {
+            // Deconstruct code into something less confusing.
+            char code[4];
+            UNMAKE_ID(code, ::LanguageIDs[i]);
 
-			// Language codes are up to three letters long.
-			code[3] = '\0';
+            // Language codes are up to three letters long.
+            code[3] = '\0';
 
-			// Try the full language code (enu).
-			loadLanguage(code, true, pass++, languageLump, len);
+            // Try the full language code (enu).
+            loadLanguage(code, true, pass++, languageLump, len);
 
-			// Try the partial language code (en).
-			code[2] = '\0';
-			loadLanguage(code, true, pass++, languageLump, len);
+            // Try the partial language code (en).
+            code[2] = '\0';
+            loadLanguage(code, true, pass++, languageLump, len);
 
-			// Try an inexact match for all languages in the same family (en_).
-			loadLanguage(code, false, pass++, languageLump, len);
-		}
-	}
+            // Try an inexact match for all languages in the same family (en_).
+            loadLanguage(code, false, pass++, languageLump, len);
+        }
+    }
 
-	// Load string defaults.
-	loadLanguage("**", true, pass++, languageLump, len);
+    // Load string defaults.
+    loadLanguage("**", true, pass++, languageLump, len);
 
-	delete[] languageLump;
+    delete[] languageLump;
 }
 
 void StringTable::prepareIndexes()
 {
-	// All of the default strings have index numbers that represent their
-	// position in the now-removed enumeration.  This function simply sets
-	// them all up.
-	for (size_t i = 0; i < ARRAY_LENGTH(::stringIndexes); i++)
-	{
-		OString name = *(::stringIndexes[i]);
-		StringHash::iterator it = _stringHash.find(name);
-		if (it == _stringHash.end())
-		{
-			TableEntry entry = {std::make_pair(false, ""), 0xFF, i};
-			_stringHash.insert(std::make_pair(name, entry));
-		}
-	}
+    // All of the default strings have index numbers that represent their
+    // position in the now-removed enumeration.  This function simply sets
+    // them all up.
+    for (size_t i = 0; i < ARRAY_LENGTH(::stringIndexes); i++)
+    {
+        OString              name = *(::stringIndexes[i]);
+        StringHash::iterator it   = _stringHash.find(name);
+        if (it == _stringHash.end())
+        {
+            TableEntry entry = {std::make_pair(false, ""), 0xFF, (int)i};
+            _stringHash.insert(std::make_pair(name, entry));
+        }
+    }
 }
 
-void StringTable::replaceEscapes(std::string& str)
+void StringTable::replaceEscapes(std::string &str)
 {
-	size_t index = 0;
+    size_t index = 0;
 
-	for (;;)
-	{
-		// Find the initial slash.
-		index = str.find("\\", index);
-		if (index == std::string::npos || index == str.length() - 1)
-			break;
+    for (;;)
+    {
+        // Find the initial slash.
+        index = str.find("\\", index);
+        if (index == std::string::npos || index == str.length() - 1)
+            break;
 
-		// Substitute the escape string.
-		switch (str.at(index + 1))
-		{
-		case 'n':
-			str.replace(index, 2, "\n");
-			break;
-		case '\\':
-			str.replace(index, 2, "\\");
-			break;
-		}
-		index += 1;
-	}
+        // Substitute the escape string.
+        switch (str.at(index + 1))
+        {
+        case 'n':
+            str.replace(index, 2, "\n");
+            break;
+        case '\\':
+            str.replace(index, 2, "\\");
+            break;
+        }
+        index += 1;
+    }
 }
 
 //
@@ -313,26 +309,26 @@ void StringTable::replaceEscapes(std::string& str)
 //
 void StringTable::dumpStrings()
 {
-	StringHash::const_iterator it = _stringHash.begin();
-	for (; it != _stringHash.end(); ++it)
-	{
-		Printf(PRINT_HIGH, "%s (pass: %d, index: %d) = %s\n", (*it).first.c_str(),
-		       (*it).second.pass, (*it).second.index, (*it).second.string.second.c_str());
-	}
+    StringHash::const_iterator it = _stringHash.begin();
+    for (; it != _stringHash.end(); ++it)
+    {
+        Printf(PRINT_HIGH, "%s (pass: %d, index: %d) = %s\n", (*it).first.c_str(), (*it).second.pass,
+               (*it).second.index, (*it).second.string.second.c_str());
+    }
 }
 
 //
 // See if a string exists in the table.
 //
-bool StringTable::hasString(const OString& name) const
+bool StringTable::hasString(const OString &name) const
 {
-	StringHash::const_iterator it = _stringHash.find(name);
-	if (it == _stringHash.end())
-		return false;
-	if ((*it).second.string.first == false)
-		return false;
+    StringHash::const_iterator it = _stringHash.find(name);
+    if (it == _stringHash.end())
+        return false;
+    if ((*it).second.string.first == false)
+        return false;
 
-	return true;
+    return true;
 }
 
 //
@@ -340,34 +336,33 @@ bool StringTable::hasString(const OString& name) const
 //
 void StringTable::loadStrings(const bool engOnly)
 {
-	clearStrings();
-	prepareIndexes();
+    clearStrings();
+    prepareIndexes();
 
-	int lump = -1;
+    int lump = -1;
 
-	lump = -1;
-	while ((lump = W_FindLump("LANGUAGE", lump)) != -1)
-	{
-		loadStringsLump(lump, "LANGUAGE", engOnly);
-	}
+    lump = -1;
+    while ((lump = W_FindLump("LANGUAGE", lump)) != -1)
+    {
+        loadStringsLump(lump, "LANGUAGE", engOnly);
+    }
 }
 
 //
 // Find a string with the same text.
 //
-const OString& StringTable::matchString(const OString& string) const
+const OString &StringTable::matchString(const OString &string) const
 {
-	for (StringHash::const_iterator it = _stringHash.begin(); it != _stringHash.end();
-	     ++it)
-	{
-		if ((*it).second.string.first == false)
-			continue;
-		if ((*it).second.string.second == string)
-			return (*it).first;
-	}
+    for (StringHash::const_iterator it = _stringHash.begin(); it != _stringHash.end(); ++it)
+    {
+        if ((*it).second.string.first == false)
+            continue;
+        if ((*it).second.string.second == string)
+            return (*it).first;
+    }
 
-	static OString empty = "";
-	return empty;
+    static OString empty = "";
+    return empty;
 }
 
 //
@@ -375,21 +370,21 @@ const OString& StringTable::matchString(const OString& string) const
 //
 // Overrides the existing string, if it exists.
 //
-void StringTable::setString(const OString& name, const OString& string)
+void StringTable::setString(const OString &name, const OString &string)
 {
-	StringHash::iterator it = _stringHash.find(name);
-	if (it == _stringHash.end())
-	{
-		// Stringtable entry does nto exist, insert it.
-		TableEntry entry = {std::make_pair(true, string), 0, -1};
-		_stringHash.insert(std::make_pair(name, entry));
-	}
-	else
-	{
-		// Stringtable entry exists, update it.
-		(*it).second.string.first = true;
-		(*it).second.string.second = string;
-	}
+    StringHash::iterator it = _stringHash.find(name);
+    if (it == _stringHash.end())
+    {
+        // Stringtable entry does nto exist, insert it.
+        TableEntry entry = {std::make_pair(true, string), 0, -1};
+        _stringHash.insert(std::make_pair(name, entry));
+    }
+    else
+    {
+        // Stringtable entry exists, update it.
+        (*it).second.string.first  = true;
+        (*it).second.string.second = string;
+    }
 }
 
 //
@@ -397,22 +392,22 @@ void StringTable::setString(const OString& name, const OString& string)
 //
 // Does not set the string if it already exists.
 //
-void StringTable::setPassString(int pass, const OString& name, const OString& string)
+void StringTable::setPassString(int pass, const OString &name, const OString &string)
 {
-	StringHash::iterator it = _stringHash.find(name);
-	if (it == _stringHash.end())
-	{
-		// Stringtable entry does not exist.
-		TableEntry entry = {std::make_pair(true, string), pass, -1};
-		_stringHash.insert(std::make_pair(name, entry));
-	}
-	else
-	{
-		// Stringtable entry exists, but has not been set yet.
-		(*it).second.string.first = true;
-		(*it).second.string.second = string;
-		(*it).second.pass = pass;
-	}
+    StringHash::iterator it = _stringHash.find(name);
+    if (it == _stringHash.end())
+    {
+        // Stringtable entry does not exist.
+        TableEntry entry = {std::make_pair(true, string), pass, -1};
+        _stringHash.insert(std::make_pair(name, entry));
+    }
+    else
+    {
+        // Stringtable entry exists, but has not been set yet.
+        (*it).second.string.first  = true;
+        (*it).second.string.second = string;
+        (*it).second.pass          = pass;
+    }
 }
 
 //
@@ -420,7 +415,7 @@ void StringTable::setPassString(int pass, const OString& name, const OString& st
 //
 size_t StringTable::size() const
 {
-	return _stringHash.size();
+    return _stringHash.size();
 }
 
 VERSION_CONTROL(stringtable_cpp, "$Id: 902a2f4d722f9ca2b833391c7d1e940f8dcc8b6c $")
